Migrate CurrencyRateController to BaseControllerV2
The CurrencyRateController has been migrated to the new base
controller. The configuration can now only be set during construction,
as it was never changed at runtime in practice with the old controller.
Similarly, the `disable` function was removed as it wasn't relied upon
in practice.

The TokenRatesController has been detached temporarily from the
CurrencyRateController, just to get this to compile and to get tests
running. This can't be merged as-is, because now the
TokenRatesController doesn't get currency rate updates anymore.